+++ server/src/main/java/org/allaymc/server/entity/effect/EffectWindChargedType.java
@@ -3,6 +3,7 @@ package org.allaymc.server.entity.effect;
 import org.allaymc.api.entity.effect.EffectInstance;
 import org.allaymc.api.entity.interfaces.EntityLiving;
 import org.allaymc.api.utils.identifier.Identifier;
+import org.allaymc.api.world.explosion.WindExplosion;
 
 import java.awt.*;
 
@@ -16,6 +17,9 @@ public class EffectWindChargedType extends AbstractEffectType {
 
     @Override
     public void onEntityDies(EntityLiving entity, EffectInstance effectInstance) {
-        //TODO: Create a wind charge explosion at the entity's position
+        var explosion = WindExplosion.playerWindCharge();
+        explosion.setSource(entity);
+        explosion.setApplySelfKnockback(false);
+        explosion.explode(entity.getDimension(), entity.getLocation());
     }
 }