+++ server/src/main/java/org/allaymc/server/entity/type/EntityTypeInitializer.java
@@ -282,4 +282,24 @@ public final class EntityTypeInitializer {
                 .addComponent(EntityPotionComponentImpl::new, EntityPotionComponentImpl.class)
                 .build();
     }
+
+    public static void initWindCharge() {
+        EntityTypes.WIND_CHARGE_PROJECTILE = AllayEntityType
+                .builder(EntityWindChargeProjectileImpl.class)
+                .vanillaEntity(EntityId.WIND_CHARGE_PROJECTILE)
+                .addComponent(EntityProjectileBaseComponentImpl::new, EntityProjectileBaseComponentImpl.class)
+                .addComponent(EntityWindChargePhysicsComponentImpl::new, EntityWindChargePhysicsComponentImpl.class)
+                .addComponent(EntityProjectileComponentImpl::new, EntityProjectileComponentImpl.class)
+                .addComponent(() -> new EntityAgeComponentImpl(), EntityAgeComponentImpl.class)
+                .build();
+
+        EntityTypes.BREEZE_WIND_CHARGE_PROJECTILE = AllayEntityType
+                .builder(EntityBreezeWindChargeProjectileImpl.class)
+                .vanillaEntity(EntityId.BREEZE_WIND_CHARGE_PROJECTILE)
+                .addComponent(EntityProjectileBaseComponentImpl::new, EntityProjectileBaseComponentImpl.class)
+                .addComponent(EntityBreezeWindChargePhysicsComponentImpl::new, EntityBreezeWindChargePhysicsComponentImpl.class)
+                .addComponent(EntityProjectileComponentImpl::new, EntityProjectileComponentImpl.class)
+                .addComponent(() -> new EntityAgeComponentImpl(), EntityAgeComponentImpl.class)
+                .build();
+    }
 }